# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event Reservation API",
        "description": "API para gestionar reservas de espacios en eventos",
        "contact": {
            "email": "soporte@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/reservations": {
            "get": {
                "summary": "Obtener todas las reservas del usuario autenticado",
                "operationId": "fbc6050d9f7fdb07075d3aaeb157ffa9",
                "responses": {
                    "200": {
                        "description": "Listado de reservas del usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "reservations": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Crear una nueva reserva",
                "operationId": "01e3d21dc330a31de025203bc142ffa3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "space_id",
                                    "event_name",
                                    "reservation_date",
                                    "start_time",
                                    "end_time"
                                ],
                                "properties": {
                                    "space_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "event_name": {
                                        "type": "string",
                                        "example": "Conferencia"
                                    },
                                    "reservation_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-10-10"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "10:00:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "12:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reserva creada con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "reservation": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "El espacio ya está reservado en ese horario"
                    }
                }
            }
        },
        "/api/reservations/{id}": {
            "get": {
                "summary": "Obtener una reserva por ID",
                "operationId": "a226b43b2a6a846245e658d01bf00720",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reserva",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserva obtenida con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "reservation": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reserva no encontrada"
                    }
                }
            },
            "put": {
                "summary": "Actualizar una reserva",
                "operationId": "02fa683a9f80a0563596fb06685bcbc7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reserva",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "event_name",
                                    "reservation_date",
                                    "start_time",
                                    "end_time"
                                ],
                                "properties": {
                                    "event_name": {
                                        "type": "string",
                                        "example": "Conferencia Actualizada"
                                    },
                                    "reservation_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-10-12"
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "09:00:00"
                                    },
                                    "end_time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "11:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reserva actualizada con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "reservation": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reserva no encontrada"
                    }
                }
            },
            "delete": {
                "summary": "Eliminar una reserva",
                "operationId": "9b6641e4ded1ff7adfce352d589b85e9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reserva",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserva eliminada con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reservation deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reserva no encontrada"
                    }
                }
            }
        },
...
    }
}